+++ module/Structure/src/Structure/Controller/Factory/VariableControllerFactory.php
@@ -6,12 +6,9 @@ use Application\Service\Role\RoleService;
 use Application\Service\Variable\VariableService;
 use Individu\Service\IndividuService;
 use Interop\Container\ContainerInterface;
-use Structure\Controller\EtablissementController;
 use Structure\Controller\VariableController;
-use Structure\Form\EtablissementForm;
 use Structure\Form\VariableForm;
 use Structure\Service\Etablissement\EtablissementService;
-use Structure\Service\Structure\StructureService;
 use Structure\Service\StructureDocument\StructureDocumentService;
 
 class VariableControllerFactory
@@ -29,15 +26,12 @@ class VariableControllerFactory
          * @var EtablissementService $etablissmentService
          * @var IndividuService $individuService
          * @var RoleService $roleService
-         * @var StructureService $structureService
          * @var StructureDocumentService $structureDocumentService
          */
         $etablissmentService = $container->get('EtablissementService');
-        $structureService = $container->get(StructureService::class);
 
         $controller = new VariableController();
         $controller->setEtablissementService($etablissmentService);
-        $controller->setStructureService($structureService);
         $controller->setVariableForm($form);
 
         $variableService = $container->get(VariableService::class);