+++ module/Structure/src/Structure/Form/Factory/VariableHydratorFactory.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Structure\Form\Factory;
+
+use Laminas\ServiceManager\Factory\FactoryInterface;
+use Psr\Container\ContainerInterface;
+use Structure\Form\Hydrator\VariableHydrator;
+
+class VariableHydratorFactory implements FactoryInterface
+{
+    public function __invoke(ContainerInterface $container, $requestedName, ?array $options = null): VariableHydrator
+    {
+        /** @var \Doctrine\ORM\EntityManager $em */
+        $em = $container->get('doctrine.entitymanager.orm_default');
+
+        return new VariableHydrator($em);
+    }
+}